backend/requirements.txt
aioredis==1.3.1
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.2
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
drf-yasg==1.17.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.23.0
service-identity==18.1.0
six==1.13.0
spotipy==2.9.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.8
zope.interface==4.7.1
